+++ contracts/lib/Bytes32Set.sol
@@ -1,4 +1,5 @@
 // SPDX-License-Identifier: BUSL-1.1
+
 pragma solidity 0.8.12;
 
 /*
@@ -22,38 +23,45 @@ THIS SOFTWARE IS NOT TESTED OR AUDITED. DO NOT USE FOR PRODUCTION.
 */
 
 library Bytes32Set {
-
     struct Set {
-        mapping(bytes32 => uint) keyPointers;
+        mapping(bytes32 => uint256) keyPointers;
         bytes32[] keyList;
     }
 
-    function insert(Set storage self, bytes32 key, string memory error) internal {
+    function insert(
+        Set storage self,
+        bytes32 key,
+        string memory error
+    ) internal {
         require(!exists(self, key), error);
         self.keyList.push(key);
         self.keyPointers[key] = self.keyList.length - 1;
     }
 
-    function remove(Set storage self, bytes32 key, string memory error) internal {
+    function remove(
+        Set storage self,
+        bytes32 key,
+        string memory error
+    ) internal {
         require(exists(self, key), error);
         bytes32 keyToMove = self.keyList[count(self) - 1];
-        uint rowToReplace = self.keyPointers[key];
+        uint256 rowToReplace = self.keyPointers[key];
         self.keyPointers[keyToMove] = rowToReplace;
         self.keyList[rowToReplace] = keyToMove;
         delete self.keyPointers[key];
         self.keyList.pop();
     }
 
-    function count(Set storage self) internal view returns(uint) {
-        return(self.keyList.length);
+    function count(Set storage self) internal view returns (uint256) {
+        return (self.keyList.length);
     }
 
-    function exists(Set storage self, bytes32 key) internal view returns(bool) {
-        if(self.keyList.length == 0) return false;
+    function exists(Set storage self, bytes32 key) internal view returns (bool) {
+        if (self.keyList.length == 0) return false;
         return self.keyList[self.keyPointers[key]] == key;
     }
 
-    function keyAtIndex(Set storage self, uint index) internal view returns(bytes32) {
+    function keyAtIndex(Set storage self, uint256 index) internal view returns (bytes32) {
         return self.keyList[index];
     }
 }
